Using StringBuilder as we are concatenating string object in a loop  using  &= operator .  use of string object in large amount of string manuplication , here string manimuplation in a loop is less efficent. String Builder is more efficient in this scenario.
String object are immutable, which means  once they are created they cannot be changed. When we use one of the methods in string class , we create a new String object

Whereas StringBuilder arre  mutable class, which means when concatenated any text to a vriable it allows us to modify the orignial value without creating new method.